changed model and view to include login required and allocated tasks sections

+++ todo/views.py
@@ -1,19 +1,23 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, get_object_or_404, redirect
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import ValidationError
 from django.http import HttpResponseRedirect
 from django.contrib import messages
-
-
 # import todo form and models
 from .forms import TodoForm
 from .models import Todo
 # Create your views here.
+
+@login_required
 def index(request):
  
     item_list = Todo.objects.order_by("-date")
     if request.method == "POST":
         form = TodoForm(request.POST)
         if form.is_valid():
-            form.save()
+            todo = form.save(commit=False)
+            todo.author = request.user
+            todo.save()
             messages.success(request,"Todo item added successfully!")
             return redirect('home')
         else:
@@ -28,11 +32,15 @@ def index(request):
     }
     return render(request, 'todo/index.html', page)
 
-    ### function to remove item, it receive todo item_id as primary key from url ##
+   
+@login_required
 def remove(request, item_id):
-    item = Todo.objects.get(id=item_id)
-    item.delete()
-    messages.info(request, "item removed !!!")
+    item = get_object_or_404(Todo, id=item_id)
+    
+    if request.user == item.author:
+        item.delete()
+        messages.info(request, "Item removed!")
+    else:
+        messages.error(request, "You do not have permission to delete this item.")
+    
     return redirect('home')
-
-
